requirements.txt
functions-framework==3.*
requests==2.*
gspread==6.*
oauth2client==4.*
google-cloud-storage